refactor(matrix): MX08 Phase 1 — split CpuMatrixBackend into src/backends/cpu-pure-ts.ts
Pure refactor.  Moves the existing CpuMatrixBackend class verbatim
from src/matrix.ts:51-73 into the new src/backends/cpu-pure-ts.ts
sibling module.  Re-exports it from matrix.ts so existing imports
(`import { CpuMatrixBackend } from "@coding-adventures/matrix"`)
keep working without source change.

Zero behaviour change

* No method signature changes.
* No timing changes (same triple-loop delegation to Matrix
  instance methods).
* All 49 existing jest tests in tests/matrix.test.ts pass without
  modification.

Why now

MX08 Phase 2 (the next PR in the chain) needs to add a Node-only
sibling at src/backends/cpu-rust-napi.ts that delegates each
MatrixBackend op into the @coding-adventures/matrix-rust-napi
addon (landed across MX07 Phases 1-4).  Splitting Phase 1 from
that landing keeps each PR mechanical and reviewable: Phase 1 is
"prove the move is a no-op"; Phase 2 is "wire up the new sibling
+ the package.json `exports` conditional that routes per
environment".

Spec sync

MX07 Phase 5 / MX08 Phase 1 row in the spec already lists this
work.  No spec edit needed in this PR — the next PR (Phase 2)
will mark Phase 1 as shipped.
